More updates for devserver
Dev server is almost fully running, we have a problem with the base route of
''.  It is not loading properly.

We should eventually:

Use dynamic imports in loadChildren, instead of strings, you might be able to skip
the module loader implementation.
This is what we use in the Angular Bazel example
https://github.com/angular/angular-bazel-example/tree/master/src